+++ type-dossier/tdt_actes/TypeDossierActesEtape.php
@@ -23,7 +23,6 @@ class TypeDossierActesEtape implements TypeDossierEtapeSetSpecificInformation
         $annulerTdtAction = $stringMapper->get('annuler-tdt');
         $tamponnerTdtAction = $stringMapper->get('tamponner-tdt');
         $teletransmissionTdtAction = $stringMapper->get('teletransmission-tdt');
-        $typologyChangeByApi = $stringMapper->get('typology-change-by-api');
         $acquiterTdtAction = $stringMapper->get('acquiter-tdt');
 
         if (!empty($typeDossierEtape->specific_type_info[self::FICHIER_ACTE])) {
@@ -34,8 +33,6 @@ class TypeDossierActesEtape implements TypeDossierEtapeSetSpecificInformation
             $result[DocumentType::ACTION][$verifTdtAction][Action::CONNECTEUR_TYPE_MAPPING][self::ARRETE]
                 = $typeDossierEtape->specific_type_info[self::FICHIER_ACTE];
             $result[DocumentType::ACTION][$tamponnerTdtAction][Action::CONNECTEUR_TYPE_MAPPING][self::ARRETE]
-                = $typeDossierEtape->specific_type_info[self::FICHIER_ACTE];
-            $result[DocumentType::ACTION][$typologyChangeByApi][Action::CONNECTEUR_TYPE_MAPPING][self::ARRETE]
                 = $typeDossierEtape->specific_type_info[self::FICHIER_ACTE];
             $result[DocumentType::THRESHOLD_SIZE] = self::THRESHOLD_SIZE;
             $result[DocumentType::THRESHOLD_FIELDS][] = $typeDossierEtape->specific_type_info[self::FICHIER_ACTE];
@@ -48,8 +45,6 @@ class TypeDossierActesEtape implements TypeDossierEtapeSetSpecificInformation
             $result[DocumentType::ACTION][$verifTdtAction][Action::CONNECTEUR_TYPE_MAPPING]
                 [self::AUTRE_DOCUMENT_ATTACHE] = $typeDossierEtape->specific_type_info[self::FICHIER_ANNEXE];
             $result[DocumentType::ACTION][$tamponnerTdtAction][Action::CONNECTEUR_TYPE_MAPPING]
-                [self::AUTRE_DOCUMENT_ATTACHE] = $typeDossierEtape->specific_type_info[self::FICHIER_ANNEXE];
-            $result[DocumentType::ACTION][$typologyChangeByApi][Action::CONNECTEUR_TYPE_MAPPING]
                 [self::AUTRE_DOCUMENT_ATTACHE] = $typeDossierEtape->specific_type_info[self::FICHIER_ANNEXE];
             $result[DocumentType::THRESHOLD_FIELDS][] = $typeDossierEtape->specific_type_info[self::FICHIER_ANNEXE];
         }